Add performance test descriptions
close #33
